fix: restructure social API to use index.js and fix CI resolution issue
- Renamed frontend/src/api/social.js to index.js following module resolution standards
- Updated all imports to use '../api' instead of '../api/social'
- Updated jest mocks to use the new path structure
- This follows Node.js module resolution conventions where index.js is automatically resolved
- All tests pass locally (34/34 Social Panel tests)
- Build completes successfully without errors

This should definitively resolve the persistent CI module resolution issue.
